Add files via upload
I recognize that this code is hideously ugly and bad and I will probably have to start over on a clean file soon. I just wanted to get as many ideas down as possible before I came to you with tons of questions. Nothing in this code is perfect at all, but I made a user menu/interface for the main and each function

There is A LOT in this file